requirements.txt

flask
gunicorn
python-dotenv
twilio
openai
requests
pydub


